Crate rusoto_stepfunctions

AWS Step Functions
AWS Step Functions is a service that lets you coordinate the components of distributed applications and microservices using visual workflows.

You can use Step Functions to build applications from individual components, each of which performs a discrete function, or task, allowing you to scale and change applications quickly. Step Functions provides a console that helps visualize the components of your application as a series of steps. Step Functions automatically triggers and tracks each step, and retries steps when there are errors, so your application executes predictably and in the right order every time. Step Functions logs the state of each step, so you can quickly diagnose and debug any issues.

Step Functions manages operations and underlying infrastructure to ensure your application is available at any scale. You can run tasks on AWS, your own servers, or any system that has access to AWS. You can access and use Step Functions using the console, the AWS SDKs, or an HTTP API. For more information about Step Functions, see the AWS Step Functions Developer Guide .

If you're using the service, you're probably looking for StepFunctionsClient and StepFunctions.

Structs

ActivityFailedEventDetails
Contains details about an activity that failed during an execution.

ActivityListItem
Contains details about an activity.

ActivityScheduleFailedEventDetails
Contains details about an activity schedule failure that occurred during an execution.

ActivityScheduledEventDetails
Contains details about an activity scheduled during an execution.

ActivityStartedEventDetails
Contains details about the start of an activity during an execution.

ActivitySucceededEventDetails
Contains details about an activity that successfully terminated during an execution.

ActivityTimedOutEventDetails
Contains details about an activity timeout that occurred during an execution.

CreateActivityInput

CreateActivityOutput

CreateStateMachineInput

CreateStateMachineOutput

DeleteActivityInput

DeleteActivityOutput

DeleteStateMachineInput

DeleteStateMachineOutput

DescribeActivityInput

DescribeActivityOutput

DescribeExecutionInput

DescribeExecutionOutput

DescribeStateMachineForExecutionInput

DescribeStateMachineForExecutionOutput

DescribeStateMachineInput

DescribeStateMachineOutput

ExecutionAbortedEventDetails
Contains details about an abort of an execution.

ExecutionFailedEventDetails
Contains details about an execution failure event.

ExecutionListItem
Contains details about an execution.

ExecutionStartedEventDetails
Contains details about the start of the execution.

ExecutionSucceededEventDetails
Contains details about the successful termination of the execution.

ExecutionTimedOutEventDetails
Contains details about the execution timeout that occurred during the execution.

GetActivityTaskInput

GetActivityTaskOutput

GetExecutionHistoryInput

GetExecutionHistoryOutput

HistoryEvent
Contains details about the events of an execution.

LambdaFunctionFailedEventDetails
Contains details about a lambda function that failed during an execution.

LambdaFunctionScheduleFailedEventDetails
Contains details about a failed lambda function schedule event that occurred during an execution.

LambdaFunctionScheduledEventDetails
Contains details about a lambda function scheduled during an execution.

LambdaFunctionStartFailedEventDetails
Contains details about a lambda function that failed to start during an execution.

LambdaFunctionSucceededEventDetails
Contains details about a lambda function that successfully terminated during an execution.

LambdaFunctionTimedOutEventDetails
Contains details about a lambda function timeout that occurred during an execution.

ListActivitiesInput

ListActivitiesOutput

ListExecutionsInput

ListExecutionsOutput

ListStateMachinesInput

ListStateMachinesOutput

ListTagsForResourceInput

ListTagsForResourceOutput

SendTaskFailureInput

SendTaskFailureOutput

SendTaskHeartbeatInput

SendTaskHeartbeatOutput

SendTaskSuccessInput

SendTaskSuccessOutput

StartExecutionInput

StartExecutionOutput

StateEnteredEventDetails
Contains details about a state entered during an execution.

StateExitedEventDetails
Contains details about an exit from a state during an execution.

StateMachineListItem
Contains details about the state machine.

StepFunctionsClient
A client for the AWS SFN API.

StopExecutionInput

StopExecutionOutput

Tag
Tags are key-value pairs that can be associated with Step Functions state machines and activities.

TagResourceInput

TagResourceOutput

TaskFailedEventDetails
Contains details about a task failure event.

TaskScheduledEventDetails
Contains details about a task scheduled during an execution.

TaskStartFailedEventDetails
Contains details about a task that failed to start during an execution.

TaskStartedEventDetails
Contains details about the start of a task during an execution.

TaskSubmitFailedEventDetails
Contains details about a task that failed to submit during an execution.

TaskSubmittedEventDetails
Contains details about a task submitted to a resource .

TaskSucceededEventDetails
Contains details about the successful completion of a task state.

TaskTimedOutEventDetails
Contains details about a resource timeout that occurred during an execution.

UntagResourceInput

UntagResourceOutput

UpdateStateMachineInput

UpdateStateMachineOutput

Enums

CreateActivityError
Errors returned by CreateActivity

CreateStateMachineError
Errors returned by CreateStateMachine

DeleteActivityError
Errors returned by DeleteActivity

DeleteStateMachineError
Errors returned by DeleteStateMachine

DescribeActivityError
Errors returned by DescribeActivity

DescribeExecutionError
Errors returned by DescribeExecution

DescribeStateMachineError
Errors returned by DescribeStateMachine

DescribeStateMachineForExecutionError
Errors returned by DescribeStateMachineForExecution

GetActivityTaskError
Errors returned by GetActivityTask

GetExecutionHistoryError
Errors returned by GetExecutionHistory

ListActivitiesError
Errors returned by ListActivities

ListExecutionsError
Errors returned by ListExecutions

ListStateMachinesError
Errors returned by ListStateMachines

ListTagsForResourceError
Errors returned by ListTagsForResource

SendTaskFailureError
Errors returned by SendTaskFailure

SendTaskHeartbeatError
Errors returned by SendTaskHeartbeat

SendTaskSuccessError
Errors returned by SendTaskSuccess

StartExecutionError
Errors returned by StartExecution

StopExecutionError
Errors returned by StopExecution

TagResourceError
Errors returned by TagResource

UntagResourceError
Errors returned by UntagResource

UpdateStateMachineError
Errors returned by UpdateStateMachine

Traits

StepFunctions
Trait representing the capabilities of the AWS SFN API. AWS SFN clients implement this trait.