Enum rusoto_stepfunctions::CreateStateMachineError

pub enum CreateStateMachineError {
    InvalidArn(String),
    InvalidDefinition(String),
    InvalidName(String),
    StateMachineAlreadyExists(String),
    StateMachineDeleting(String),
    StateMachineLimitExceeded(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by CreateStateMachine

Variants
InvalidArn(String)

The provided Amazon Resource Name (ARN) is invalid.

InvalidDefinition(String)

The provided Amazon States Language definition is invalid.

InvalidName(String)

The provided name is invalid.

StateMachineAlreadyExists(String)

A state machine with the same name but a different definition or role ARN already exists.

StateMachineDeleting(String)

The specified state machine is being deleted.

StateMachineLimitExceeded(String)

The maximum number of state machines has been reached. Existing state machines must be deleted before a new state machine can be created.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl CreateStateMachineError

fn from_body(body: &str) -> CreateStateMachineError