Enum rusoto_stepfunctions::StartExecutionError

pub enum StartExecutionError {
    ExecutionAlreadyExists(String),
    ExecutionLimitExceeded(String),
    InvalidArn(String),
    InvalidExecutionInput(String),
    InvalidName(String),
    StateMachineDeleting(String),
    StateMachineDoesNotExist(String),
    HttpDispatch(HttpDispatchError),
    Credentials(CredentialsError),
    Validation(String),
    Unknown(String),
}

Errors returned by StartExecution

Variants
ExecutionAlreadyExists(String)

The execution has the same name as another execution (but a different input).

Executions with the same name and input are considered idempotent.

ExecutionLimitExceeded(String)

The maximum number of running executions has been reached. Running executions must end or be stopped before a new execution can be started.

InvalidArn(String)

The provided Amazon Resource Name (ARN) is invalid.

InvalidExecutionInput(String)

The provided JSON input data is invalid.

InvalidName(String)

The provided name is invalid.

StateMachineDeleting(String)

The specified state machine is being deleted.

StateMachineDoesNotExist(String)

The specified state machine does not exist.

HttpDispatch(HttpDispatchError)

An error occurred dispatching the HTTP request

Credentials(CredentialsError)

An error was encountered with AWS credentials.

Validation(String)

A validation error occurred. Details from AWS are provided.

Unknown(String)

An unknown error occurred. The raw HTTP response is provided.

Methods

impl StartExecutionError

pub fn from_body(body: &str) -> StartExecutionError