Struct rusoto_stepfunctions::HistoryEvent

pub struct HistoryEvent {
    pub activity_failed_event_details: Option<ActivityFailedEventDetails>,
    pub activity_schedule_failed_event_details: Option<ActivityScheduleFailedEventDetails>,
    pub activity_scheduled_event_details: Option<ActivityScheduledEventDetails>,
    pub activity_started_event_details: Option<ActivityStartedEventDetails>,
    pub activity_succeeded_event_details: Option<ActivitySucceededEventDetails>,
    pub activity_timed_out_event_details: Option<ActivityTimedOutEventDetails>,
    pub execution_aborted_event_details: Option<ExecutionAbortedEventDetails>,
    pub execution_failed_event_details: Option<ExecutionFailedEventDetails>,
    pub execution_started_event_details: Option<ExecutionStartedEventDetails>,
    pub execution_succeeded_event_details: Option<ExecutionSucceededEventDetails>,
    pub execution_timed_out_event_details: Option<ExecutionTimedOutEventDetails>,
    pub id: i64,
    pub lambda_function_failed_event_details: Option<LambdaFunctionFailedEventDetails>,
    pub lambda_function_schedule_failed_event_details: Option<LambdaFunctionScheduleFailedEventDetails>,
    pub lambda_function_scheduled_event_details: Option<LambdaFunctionScheduledEventDetails>,
    pub lambda_function_start_failed_event_details: Option<LambdaFunctionStartFailedEventDetails>,
    pub lambda_function_succeeded_event_details: Option<LambdaFunctionSucceededEventDetails>,
    pub lambda_function_timed_out_event_details: Option<LambdaFunctionTimedOutEventDetails>,
    pub previous_event_id: Option<i64>,
    pub state_entered_event_details: Option<StateEnteredEventDetails>,
    pub state_exited_event_details: Option<StateExitedEventDetails>,
    pub task_failed_event_details: Option<TaskFailedEventDetails>,
    pub task_scheduled_event_details: Option<TaskScheduledEventDetails>,
    pub task_start_failed_event_details: Option<TaskStartFailedEventDetails>,
    pub task_started_event_details: Option<TaskStartedEventDetails>,
    pub task_submit_failed_event_details: Option<TaskSubmitFailedEventDetails>,
    pub task_submitted_event_details: Option<TaskSubmittedEventDetails>,
    pub task_succeeded_event_details: Option<TaskSucceededEventDetails>,
    pub task_timed_out_event_details: Option<TaskTimedOutEventDetails>,
    pub timestamp: f64,
    pub type_: String,
}

Contains details about the events of an execution.

Fields
activity_failed_event_details: Option<ActivityFailedEventDetails>activity_schedule_failed_event_details: Option<ActivityScheduleFailedEventDetails>

Contains details about an activity schedule event that failed during an execution.

activity_scheduled_event_details: Option<ActivityScheduledEventDetails>activity_started_event_details: Option<ActivityStartedEventDetails>activity_succeeded_event_details: Option<ActivitySucceededEventDetails>activity_timed_out_event_details: Option<ActivityTimedOutEventDetails>execution_aborted_event_details: Option<ExecutionAbortedEventDetails>execution_failed_event_details: Option<ExecutionFailedEventDetails>execution_started_event_details: Option<ExecutionStartedEventDetails>execution_succeeded_event_details: Option<ExecutionSucceededEventDetails>execution_timed_out_event_details: Option<ExecutionTimedOutEventDetails>id: i64

The id of the event. Events are numbered sequentially, starting at one.

lambda_function_failed_event_details: Option<LambdaFunctionFailedEventDetails>lambda_function_schedule_failed_event_details: Option<LambdaFunctionScheduleFailedEventDetails>lambda_function_scheduled_event_details: Option<LambdaFunctionScheduledEventDetails>lambda_function_start_failed_event_details: Option<LambdaFunctionStartFailedEventDetails>

Contains details about a lambda function that failed to start during an execution.

lambda_function_succeeded_event_details: Option<LambdaFunctionSucceededEventDetails>

Contains details about a lambda function that terminated successfully during an execution.

lambda_function_timed_out_event_details: Option<LambdaFunctionTimedOutEventDetails>previous_event_id: Option<i64>

The id of the previous event.

state_entered_event_details: Option<StateEnteredEventDetails>state_exited_event_details: Option<StateExitedEventDetails>task_failed_event_details: Option<TaskFailedEventDetails>

Contains details about the failure of a task.

task_scheduled_event_details: Option<TaskScheduledEventDetails>

Contains details about a task that was scheduled.

task_start_failed_event_details: Option<TaskStartFailedEventDetails>

Contains details about a task that failed to start.

task_started_event_details: Option<TaskStartedEventDetails>

Contains details about a task that was started.

task_submit_failed_event_details: Option<TaskSubmitFailedEventDetails>

Contains details about a task that where the submit failed.

task_submitted_event_details: Option<TaskSubmittedEventDetails>

Contains details about a submitted task.

task_succeeded_event_details: Option<TaskSucceededEventDetails>

Contains details about a task that succeeded.

task_timed_out_event_details: Option<TaskTimedOutEventDetails>

Contains details about a task that timed out.

timestamp: f64

The date and time the event occurred.

type_: String

The type of the event.