Trait rusoto_stepfunctions::StepFunctions

pub trait StepFunctions {
    fn create_activity(
        &self, 
        input: &CreateActivityInput
    ) -> Result<CreateActivityOutput, CreateActivityError>;
    fn create_state_machine(
        &self, 
        input: &CreateStateMachineInput
    ) -> Result<CreateStateMachineOutput, CreateStateMachineError>;
    fn delete_activity(
        &self, 
        input: &DeleteActivityInput
    ) -> Result<DeleteActivityOutput, DeleteActivityError>;
    fn delete_state_machine(
        &self, 
        input: &DeleteStateMachineInput
    ) -> Result<DeleteStateMachineOutput, DeleteStateMachineError>;
    fn describe_activity(
        &self, 
        input: &DescribeActivityInput
    ) -> Result<DescribeActivityOutput, DescribeActivityError>;
    fn describe_execution(
        &self, 
        input: &DescribeExecutionInput
    ) -> Result<DescribeExecutionOutput, DescribeExecutionError>;
    fn describe_state_machine(
        &self, 
        input: &DescribeStateMachineInput
    ) -> Result<DescribeStateMachineOutput, DescribeStateMachineError>;
    fn get_activity_task(
        &self, 
        input: &GetActivityTaskInput
    ) -> Result<GetActivityTaskOutput, GetActivityTaskError>;
    fn get_execution_history(
        &self, 
        input: &GetExecutionHistoryInput
    ) -> Result<GetExecutionHistoryOutput, GetExecutionHistoryError>;
    fn list_activities(
        &self, 
        input: &ListActivitiesInput
    ) -> Result<ListActivitiesOutput, ListActivitiesError>;
    fn list_executions(
        &self, 
        input: &ListExecutionsInput
    ) -> Result<ListExecutionsOutput, ListExecutionsError>;
    fn list_state_machines(
        &self, 
        input: &ListStateMachinesInput
    ) -> Result<ListStateMachinesOutput, ListStateMachinesError>;
    fn send_task_failure(
        &self, 
        input: &SendTaskFailureInput
    ) -> Result<SendTaskFailureOutput, SendTaskFailureError>;
    fn send_task_heartbeat(
        &self, 
        input: &SendTaskHeartbeatInput
    ) -> Result<SendTaskHeartbeatOutput, SendTaskHeartbeatError>;
    fn send_task_success(
        &self, 
        input: &SendTaskSuccessInput
    ) -> Result<SendTaskSuccessOutput, SendTaskSuccessError>;
    fn start_execution(
        &self, 
        input: &StartExecutionInput
    ) -> Result<StartExecutionOutput, StartExecutionError>;
    fn stop_execution(
        &self, 
        input: &StopExecutionInput
    ) -> Result<StopExecutionOutput, StopExecutionError>;
}

Trait representing the capabilities of the AWS SFN API. AWS SFN clients implement this trait.

Required Methods

fn create_activity(
&self,
input: &CreateActivityInput
) -> Result<CreateActivityOutput, CreateActivityError>

Creates an activity.

fn create_state_machine(
&self,
input: &CreateStateMachineInput
) -> Result<CreateStateMachineOutput, CreateStateMachineError>

Creates a state machine.

fn delete_activity(
&self,
input: &DeleteActivityInput
) -> Result<DeleteActivityOutput, DeleteActivityError>

Deletes an activity.

fn delete_state_machine(
&self,
input: &DeleteStateMachineInput
) -> Result<DeleteStateMachineOutput, DeleteStateMachineError>

Deletes a state machine. This is an asynchronous operation-- it sets the state machine's status to "DELETING" and begins the delete process.

fn describe_activity(
&self,
input: &DescribeActivityInput
) -> Result<DescribeActivityOutput, DescribeActivityError>

Describes an activity.

fn describe_execution(
&self,
input: &DescribeExecutionInput
) -> Result<DescribeExecutionOutput, DescribeExecutionError>

Describes an execution.

fn describe_state_machine(
&self,
input: &DescribeStateMachineInput
) -> Result<DescribeStateMachineOutput, DescribeStateMachineError>

Describes a state machine.

fn get_activity_task(
&self,
input: &GetActivityTaskInput
) -> Result<GetActivityTaskOutput, GetActivityTaskError>

Used by workers to retrieve a task (with the specified activity ARN) scheduled for execution by a running state machine. This initiates a long poll, where the service holds the HTTP connection open and responds as soon as a task becomes available (i.e. an execution of a task of this type is needed.) The maximum time the service holds on to the request before responding is 60 seconds. If no task is available within 60 seconds, the poll will return an empty result, that is, the taskToken returned is an empty string.

Workers should set their client side socket timeout to at least 65 seconds (5 seconds higher than the maximum time the service may hold the poll request).

fn get_execution_history(
&self,
input: &GetExecutionHistoryInput
) -> Result<GetExecutionHistoryOutput, GetExecutionHistoryError>

Returns the history of the specified execution as a list of events. By default, the results are returned in ascending order of the timeStamp of the events. Use the reverseOrder parameter to get the latest events first. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextToken returned by the previous call.

fn list_activities(
&self,
input: &ListActivitiesInput
) -> Result<ListActivitiesOutput, ListActivitiesError>

Lists the existing activities. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextToken returned by the previous call.

fn list_executions(
&self,
input: &ListExecutionsInput
) -> Result<ListExecutionsOutput, ListExecutionsError>

Lists the executions of a state machine that meet the filtering criteria. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextToken returned by the previous call.

fn list_state_machines(
&self,
input: &ListStateMachinesInput
) -> Result<ListStateMachinesOutput, ListStateMachinesError>

Lists the existing state machines. The results may be split into multiple pages. To retrieve subsequent pages, make the call again using the nextToken returned by the previous call.

fn send_task_failure(
&self,
input: &SendTaskFailureInput
) -> Result<SendTaskFailureOutput, SendTaskFailureError>

Used by workers to report that the task identified by the taskToken failed.

fn send_task_heartbeat(
&self,
input: &SendTaskHeartbeatInput
) -> Result<SendTaskHeartbeatOutput, SendTaskHeartbeatError>

Used by workers to report to the service that the task represented by the specified taskToken is still making progress. This action resets the Heartbeat clock. The Heartbeat threshold is specified in the state machine's Amazon States Language definition. This action does not in itself create an event in the execution history. However, if the task times out, the execution history will contain an ActivityTimedOut event.

The Timeout of a task, defined in the state machine's Amazon States Language definition, is its maximum allowed duration, regardless of the number of SendTaskHeartbeat requests received.

This operation is only useful for long-lived tasks to report the liveliness of the task.

fn send_task_success(
&self,
input: &SendTaskSuccessInput
) -> Result<SendTaskSuccessOutput, SendTaskSuccessError>

Used by workers to report that the task identified by the taskToken completed successfully.

fn start_execution(
&self,
input: &StartExecutionInput
) -> Result<StartExecutionOutput, StartExecutionError>

Starts a state machine execution.

fn stop_execution(
&self,
input: &StopExecutionInput
) -> Result<StopExecutionOutput, StopExecutionError>

Stops an execution.