Struct rusoto_stepfunctions::StepFunctionsClient

pub struct StepFunctionsClient { /* fields omitted */ }

A client for the AWS SFN API.

Methods

impl StepFunctionsClient

pub fn new(region: Region) -> StepFunctionsClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> StepFunctionsClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,