Trait rusoto_stepfunctions::StepFunctions

pub trait StepFunctions {
    fn create_activity(
        &self, 
        input: CreateActivityInput
    ) -> RusotoFuture<CreateActivityOutput, CreateActivityError>;

    fn create_state_machine(
        &self, 
        input: CreateStateMachineInput
    ) -> RusotoFuture<CreateStateMachineOutput, CreateStateMachineError>;

    fn delete_activity(
        &self, 
        input: DeleteActivityInput
    ) -> RusotoFuture<DeleteActivityOutput, DeleteActivityError>;

    fn delete_state_machine(
        &self, 
        input: DeleteStateMachineInput
    ) -> RusotoFuture<DeleteStateMachineOutput, DeleteStateMachineError>;

    fn describe_activity(
        &self, 
        input: DescribeActivityInput
    ) -> RusotoFuture<DescribeActivityOutput, DescribeActivityError>;

    fn describe_execution(
        &self, 
        input: DescribeExecutionInput
    ) -> RusotoFuture<DescribeExecutionOutput, DescribeExecutionError>;

    fn describe_state_machine(
        &self, 
        input: DescribeStateMachineInput
    ) -> RusotoFuture<DescribeStateMachineOutput, DescribeStateMachineError>;

    fn describe_state_machine_for_execution(
        &self, 
        input: DescribeStateMachineForExecutionInput
    ) -> RusotoFuture<DescribeStateMachineForExecutionOutput, DescribeStateMachineForExecutionError>;

    fn get_activity_task(
        &self, 
        input: GetActivityTaskInput
    ) -> RusotoFuture<GetActivityTaskOutput, GetActivityTaskError>;

    fn get_execution_history(
        &self, 
        input: GetExecutionHistoryInput
    ) -> RusotoFuture<GetExecutionHistoryOutput, GetExecutionHistoryError>;

    fn list_activities(
        &self, 
        input: ListActivitiesInput
    ) -> RusotoFuture<ListActivitiesOutput, ListActivitiesError>;

    fn list_executions(
        &self, 
        input: ListExecutionsInput
    ) -> RusotoFuture<ListExecutionsOutput, ListExecutionsError>;

    fn list_state_machines(
        &self, 
        input: ListStateMachinesInput
    ) -> RusotoFuture<ListStateMachinesOutput, ListStateMachinesError>;

    fn list_tags_for_resource(
        &self, 
        input: ListTagsForResourceInput
    ) -> RusotoFuture<ListTagsForResourceOutput, ListTagsForResourceError>;

    fn send_task_failure(
        &self, 
        input: SendTaskFailureInput
    ) -> RusotoFuture<SendTaskFailureOutput, SendTaskFailureError>;

    fn send_task_heartbeat(
        &self, 
        input: SendTaskHeartbeatInput
    ) -> RusotoFuture<SendTaskHeartbeatOutput, SendTaskHeartbeatError>;

    fn send_task_success(
        &self, 
        input: SendTaskSuccessInput
    ) -> RusotoFuture<SendTaskSuccessOutput, SendTaskSuccessError>;

    fn start_execution(
        &self, 
        input: StartExecutionInput
    ) -> RusotoFuture<StartExecutionOutput, StartExecutionError>;

    fn stop_execution(
        &self, 
        input: StopExecutionInput
    ) -> RusotoFuture<StopExecutionOutput, StopExecutionError>;

    fn tag_resource(
        &self, 
        input: TagResourceInput
    ) -> RusotoFuture<TagResourceOutput, TagResourceError>;

    fn untag_resource(
        &self, 
        input: UntagResourceInput
    ) -> RusotoFuture<UntagResourceOutput, UntagResourceError>;

    fn update_state_machine(
        &self, 
        input: UpdateStateMachineInput
    ) -> RusotoFuture<UpdateStateMachineOutput, UpdateStateMachineError>;
}

Trait representing the capabilities of the AWS SFN API. AWS SFN clients implement this trait.

Required methods

fn create_activity(
&self,
input: CreateActivityInput
) -> RusotoFuture<CreateActivityOutput, CreateActivityError>

Creates an activity. An activity is a task that you write in any programming language and host on any machine that has access to AWS Step Functions. Activities must poll Step Functions using the GetActivityTask API action and respond using SendTask* API actions. This function lets Step Functions know the existence of your activity and returns an identifier for use in a state machine and when polling from the activity.

This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.

fn create_state_machine(
&self,
input: CreateStateMachineInput
) -> RusotoFuture<CreateStateMachineOutput, CreateStateMachineError>

Creates a state machine. A state machine consists of a collection of states that can do work (Task states), determine to which states to transition next (Choice states), stop an execution with an error (Fail states), and so on. State machines are specified using a JSON-based, structured language.

This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.

fn delete_activity(
&self,
input: DeleteActivityInput
) -> RusotoFuture<DeleteActivityOutput, DeleteActivityError>

Deletes an activity.

fn delete_state_machine(
&self,
input: DeleteStateMachineInput
) -> RusotoFuture<DeleteStateMachineOutput, DeleteStateMachineError>

Deletes a state machine. This is an asynchronous operation: It sets the state machine's status to DELETING and begins the deletion process. Each state machine execution is deleted the next time it makes a state transition.

The state machine itself is deleted after all executions are completed or deleted.

fn describe_activity(
&self,
input: DescribeActivityInput
) -> RusotoFuture<DescribeActivityOutput, DescribeActivityError>

Describes an activity.

This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.

fn describe_execution(
&self,
input: DescribeExecutionInput
) -> RusotoFuture<DescribeExecutionOutput, DescribeExecutionError>

Describes an execution.

This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.

fn describe_state_machine(
&self,
input: DescribeStateMachineInput
) -> RusotoFuture<DescribeStateMachineOutput, DescribeStateMachineError>

Describes a state machine.

This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.

fn describe_state_machine_for_execution(
&self,
input: DescribeStateMachineForExecutionInput
) -> RusotoFuture<DescribeStateMachineForExecutionOutput, DescribeStateMachineForExecutionError>

Describes the state machine associated with a specific execution.

This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.

fn get_activity_task(
&self,
input: GetActivityTaskInput
) -> RusotoFuture<GetActivityTaskOutput, GetActivityTaskError>

Used by workers to retrieve a task (with the specified activity ARN) which has been scheduled for execution by a running state machine. This initiates a long poll, where the service holds the HTTP connection open and responds as soon as a task becomes available (i.e. an execution of a task of this type is needed.) The maximum time the service holds on to the request before responding is 60 seconds. If no task is available within 60 seconds, the poll returns a taskToken with a null string.

Workers should set their client side socket timeout to at least 65 seconds (5 seconds higher than the maximum time the service may hold the poll request).

Polling with GetActivityTask can cause latency in some implementations. See Avoid Latency When Polling for Activity Tasks in the Step Functions Developer Guide.

fn get_execution_history(
&self,
input: GetExecutionHistoryInput
) -> RusotoFuture<GetExecutionHistoryOutput, GetExecutionHistoryError>

Returns the history of the specified execution as a list of events. By default, the results are returned in ascending order of the timeStamp of the events. Use the reverseOrder parameter to get the latest events first.

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

fn list_activities(
&self,
input: ListActivitiesInput
) -> RusotoFuture<ListActivitiesOutput, ListActivitiesError>

Lists the existing activities.

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.

fn list_executions(
&self,
input: ListExecutionsInput
) -> RusotoFuture<ListExecutionsOutput, ListExecutionsError>

Lists the executions of a state machine that meet the filtering criteria. Results are sorted by time, with the most recent execution first.

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.

fn list_state_machines(
&self,
input: ListStateMachinesInput
) -> RusotoFuture<ListStateMachinesOutput, ListStateMachinesError>

Lists the existing state machines.

If nextToken is returned, there are more results available. The value of nextToken is a unique pagination token for each page. Make the call again using the returned token to retrieve the next page. Keep all other arguments unchanged. Each pagination token expires after 24 hours. Using an expired pagination token will return an HTTP 400 InvalidToken error.

This operation is eventually consistent. The results are best effort and may not reflect very recent updates and changes.

fn list_tags_for_resource(
&self,
input: ListTagsForResourceInput
) -> RusotoFuture<ListTagsForResourceOutput, ListTagsForResourceError>

List tags for a given resource.

fn send_task_failure(
&self,
input: SendTaskFailureInput
) -> RusotoFuture<SendTaskFailureOutput, SendTaskFailureError>

Used by workers to report that the task identified by the taskToken failed.

fn send_task_heartbeat(
&self,
input: SendTaskHeartbeatInput
) -> RusotoFuture<SendTaskHeartbeatOutput, SendTaskHeartbeatError>

Used by workers to report to the service that the task represented by the specified taskToken is still making progress. This action resets the Heartbeat clock. The Heartbeat threshold is specified in the state machine's Amazon States Language definition. This action does not in itself create an event in the execution history. However, if the task times out, the execution history contains an ActivityTimedOut event.

The Timeout of a task, defined in the state machine's Amazon States Language definition, is its maximum allowed duration, regardless of the number of SendTaskHeartbeat requests received.

This operation is only useful for long-lived tasks to report the liveliness of the task.

fn send_task_success(
&self,
input: SendTaskSuccessInput
) -> RusotoFuture<SendTaskSuccessOutput, SendTaskSuccessError>

Used by workers to report that the task identified by the taskToken completed successfully.

fn start_execution(
&self,
input: StartExecutionInput
) -> RusotoFuture<StartExecutionOutput, StartExecutionError>

Starts a state machine execution.

StartExecution is idempotent. If StartExecution is called with the same name and input as a running execution, the call will succeed and return the same response as the original request. If the execution is closed or if the input is different, it will return a 400 ExecutionAlreadyExists error. Names can be reused after 90 days.

fn stop_execution(
&self,
input: StopExecutionInput
) -> RusotoFuture<StopExecutionOutput, StopExecutionError>

Stops an execution.

fn tag_resource(
&self,
input: TagResourceInput
) -> RusotoFuture<TagResourceOutput, TagResourceError>

Add a tag to a Step Functions resource.

fn untag_resource(
&self,
input: UntagResourceInput
) -> RusotoFuture<UntagResourceOutput, UntagResourceError>

Remove a tag from a Step Functions resource

fn update_state_machine(
&self,
input: UpdateStateMachineInput
) -> RusotoFuture<UpdateStateMachineOutput, UpdateStateMachineError>

Updates an existing state machine by modifying its definition and/or roleArn. Running executions will continue to use the previous definition and roleArn. You must include at least one of definition or roleArn or you will receive a MissingRequiredParameter error.

All StartExecution calls within a few seconds will use the updated definition and roleArn. Executions started immediately after calling UpdateStateMachine may use the previous state machine definition and roleArn.

Loading content...